Split package module further to avoid recursive imports
This completes the previous modularization effort by splitting into
constants, bciclasses and utils.
