+++ src/bci_build/package/node.py
@@ -2,21 +2,21 @@
 import datetime
 from typing import Literal
 
-from bci_build.package import _SUPPORTED_UNTIL_SLE
-from bci_build.package import LanguageStackContainer
-from bci_build.package import OsVersion
-from bci_build.package import SupportLevel
+from bci_build.package.bciclasses import LanguageStackContainer
 
+from .constants import OsVersion
+from .constants import SUPPORTED_UNTIL_SLE
+from .constants import SupportLevel
 
 # see https://raw.githubusercontent.com/nodejs/Release/main/README.md
 _NODEJS_SUPPORT_ENDS = {
     20: datetime.date(2026, 4, 30),
     # ... upstream is 2024/4/30 but SUSE ends earlier with SP5
     # see https://confluence.suse.com/display/SLE/Node.js
-    18: _SUPPORTED_UNTIL_SLE[OsVersion.SP5],
+    18: SUPPORTED_UNTIL_SLE[OsVersion.SP5],
     # upstream 2023/9/11 but SUSE extends end of general support SP4
     # see https://confluence.suse.com/display/SLE/Node.js
-    16: _SUPPORTED_UNTIL_SLE[OsVersion.SP4],
+    16: SUPPORTED_UNTIL_SLE[OsVersion.SP4],
 }
