+++ src/bci_build/package/python.py
@@ -1,13 +1,15 @@
-"""Python language related BCI containers"""
+"""This module contains classes and functions for generating Dockerfiles for Python language related BCI containers.
+"""
 import datetime
 from typing import Literal
 
-from bci_build.package import _SUPPORTED_UNTIL_SLE
-from bci_build.package import CAN_BE_LATEST_OS_VERSION
-from bci_build.package import LanguageStackContainer
-from bci_build.package import OsVersion
 from bci_build.package import Replacement
-from bci_build.package import SupportLevel
+from bci_build.package.bciclasses import LanguageStackContainer
+
+from .constants import CAN_BE_LATEST_OS_VERSION
+from .constants import OsVersion
+from .constants import SUPPORTED_UNTIL_SLE
+from .constants import SupportLevel
 
 
 def _get_python_kwargs(
@@ -89,7 +91,7 @@ PYTHON_3_10_SP4 = LanguageStackContainer(
     support_level=SupportLevel.L3,
     is_latest=False,
     **_get_python_kwargs("3.10", OsVersion.SP4),
-    supported_until=_SUPPORTED_UNTIL_SLE[OsVersion.SP4],
+    supported_until=SUPPORTED_UNTIL_SLE[OsVersion.SP4],
 )
 
 PYTHON_3_11_CONTAINERS = (
